Very very fisrt version.
It was converted from that repository 'scahp.github.io' which is javascript version of shadows.
To improve performance (by using gs and massive file read fast from files) it is ported to here.
This is very first version. so I should do refactoring for this. It will be done soon.